infrastructure/bedrock/agentcore/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
GitPython
